LayoutTests/imported/w3c/web-platform-tests/scroll-animations/css/get-animations-inactive-timeline-expected.txt


FAIL getAnimations includes inactive scroll-linked animations that have not been canceled assert_equals: Inactive timeline when timeline's source element cannot be scrolled expected null but got object "0%"
